src/requirements.txt
Flask==2.3.3
Flask-Session>=0.5.0
python-dotenv==1.0.0
spacy==3.7.5
PyPDF2==3.0.1
python-docx==1.2.0
scikit-learn==1.3.2
numpy==1.23.5
sentence-transformers==2.2.2
pandas==2.1.4
spacy-lookups-data==1.0.5
transformers<4.35
tokenizers<0.15
google-generativeai>=0.3.0
langdetect>=1.0.9
googletrans==4.0.0rc1
openpyxl==3.1.2
reportlab==4.0.7